# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OtpSendInput {
  /** Email address to verify */
  email: string;
}

export interface OtpSendResponse {
  success: boolean;
  /** Approximate verification code lifetime in seconds */
  expiresIn: number;
}

export interface OtpVerifyInput {
  email: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
}

export interface OtpVerifyResponse {
  verified: boolean;
}