# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Email verification operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendInput"
      responses:
        "200":
          description: Verification code accepted for delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResponse"
        "400":
          description: Invalid email address
        "502":
          description: Email provider error
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid code or email address
        "502":
          description: Email provider error
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OtpSendInput:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Email address to verify
      required:
        - email
    OtpSendResponse:
      type: object
      properties:
        success:
          type: boolean
        expiresIn:
          type: integer
          description: Approximate verification code lifetime in seconds
      required:
        - success
        - expiresIn
    OtpVerifyInput:
      type: object
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: "^[0-9]{6}$"
      required:
        - email
        - code
    OtpVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
      required:
        - verified