# lib/api-zod/src/generated/types/otpVerifyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OtpVerifyInput {
  email: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
}